3 курс/5 семестр/Разработка баз данных/Интерфейс/requirements.txt
mysql-connector-python
tk
